Move git state/actions into a component
